+++ src/pages/demos/charge-demo.tsx
@@ -1,0 +1,10 @@
+import { Charge } from '@/components';
+import styles from './style/index.module.less';
+
+const ChargeDemo = () => (
+  <div className={styles.chargeDemo}>
+    <Charge />
+  </div>
+);
+
+export default ChargeDemo;